Agent-RakFort/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
pandas 